Start porting Blueprints v2

+++ packages/playground/common/src/index.ts
@@ -75,9 +75,10 @@ export const unzipFile = async (
 
 export const zipDirectory = async (
 	php: UniversalPHP,
-	directoryPath: string
+	directoryPath: string,
+	zipPath?: string
 ) => {
-	const outputPath = `/tmp/file${Math.random()}.zip`;
+	const outputPath = zipPath || `/tmp/file${Math.random()}.zip`;
 	const js = phpVars({
 		directoryPath,
 		outputPath,
@@ -106,6 +107,10 @@ export const zipDirectory = async (
 		zipDirectory(${js.directoryPath}, ${js.outputPath});
 		`,
 	});
+
+	if (zipPath) {
+		return undefined;
+	}
 
 	const fileBuffer = await php.readFileAsBuffer(outputPath);
 	php.unlink(outputPath);
